requirements.txt
absl-py==2.1.0
anytree==2.10.0
astunparse==1.6.3
beautifulsoup4==4.12.0
blinker==1.8.2
bokeh==3.4.1
Bottleneck==1.3.4
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
cgen==2020.1
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
codepy==2023.1
colorama==0.4.6
contourpy==1.2.1
cryptography==3.4.8
cycler==0.10.0
Cython==3.0.5
dask==2024.5.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
devito==4.8.3
eofs==1.4.1
et-xmlfile==1.1.0
fastjsonschema==2.16.2
filelock==3.12.2
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.25.0
fsspec==2023.4.0
future==0.18.3
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.63.0
h5py==3.11.0
idna==3.4
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
libclang==18.1.1
livelossplot==0.5.5
llvmlite==0.41.1
locket==1.0.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
metakernel==0.30.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.3.2
mpltools==0.2.0
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
namex==0.0.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.0
octave_kernel==0.36.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
pandas==2.1.1
pandocfilters==1.4.3
parso==0.8.3
partd==1.4.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycparser==2.21
pydot==1.4.2
Pygments==2.15.1
pynndescent==0.5.10
pyOpenSSL==22.1.0
pyparsing==3.1.0
PyQt5==5.15.10
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.1
pytools==2023.1.1
pytz==2023.3
pywin32==305.1
PyYAML==6.0
pyzmq==25.0.2
QtPy==2.4.0
regex==2023.10.3
requests==2.31.0
rich==13.7.1
rpds-py==0.12.2
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==2.2.0
tinycss2==1.2.1
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torcheval==0.0.7
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
umap-learn==0.5.5
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.3
wrapt==1.14.1
xlrd==2.0.1
xyzservices==2024.4.0
y-py==0.6.0
ypy-websocket==0.7.1
zipp==3.18.1
gunicorn==20.1.0
